new file:   Doxyfile 	deleted:    LCC.so 	modified:   LCC_gravityby1.c 	modified:   LCF.c 	deleted:    LCF.so 	new file:   LCF_pleph.c 	renamed:    lua/lauxlib.h -> api/lauxlib.h 	renamed:    lua/liblua.a -> api/liblua.a 	renamed:    lua/lua.h -> api/lua.h 	renamed:    lua/lua.hpp -> api/lua.hpp 	renamed:    lua/luaconf.h -> api/luaconf.h 	renamed:    lua/lualib.h -> api/lualib.h 	renamed:    lua/lua -> lua 	modified:   makefile 	new file:   read 	modified:   read.c         new directory doxygen

+++ SSM/read.c
@@ -2,7 +2,7 @@
 
 void main()
 {
-  int jdate;
+  double  jdate;
   int     planet,center; 
   double  state[6];
